Completed object selection.
Objects can now be made selectable by adding the SelectableObject class to them. The square in the bottom-left of the screen will change to the sprite color of a selected object (this is temporary and will be replaced with the object's sprite once we have sprites). The Clear Selection button clears the current selection when pressed, and the player can only select one object at a time.