Finished 130 task
The player can transfer food that is on the plate into the Cafeteria scene and back to the Kitchen scene. When the food is placed on a plate the Guests update to whether that was their order or not. If it is their order they "take" it and the food is deleted.